Weather Forecast needed split up to conform to standards
