+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/WeatherForecast/WeatherForecastState.Behavior.cs
@@ -1,0 +1,20 @@
+﻿namespace BlazorHosted_CSharp.Client.Features.WeatherForecast
+{
+  using BlazorHosted_CSharp.Shared.Features.WeatherForecast;
+  using BlazorState;
+  
+  public partial class WeatherForecastsState : State<WeatherForecastsState>
+  {
+    private WeatherForecastsState(WeatherForecastsState aState) : this()
+    {
+      foreach (WeatherForecastDto weatherForecastDto in aState.WeatherForecasts)
+      {
+        _WeatherForecasts.Add(weatherForecastDto.Clone() as WeatherForecastDto);
+      }
+    }
+
+    public override object Clone() => new WeatherForecastsState(this);
+
+    protected override void Initialize() { }
+  }
+}
