+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/WeatherForecast/WeatherForecastState.Debug.cs
@@ -1,6 +1,7 @@
 ﻿namespace BlazorHosted_CSharp.Client.Features.WeatherForecast
 {
   using System.Collections.Generic;
+  using System.Reflection;
   using BlazorHosted_CSharp.Shared.Features.WeatherForecast;
   using BlazorState;
   using Microsoft.JSInterop;
@@ -17,6 +18,13 @@
       };
 
       return newWeatherForecastsState;
+    }
+
+    private void Initialize(List<WeatherForecastDto> aWeatherForecastList)
+    {
+      ThrowIfNotTestAssembly(Assembly.GetCallingAssembly());
+      _WeatherForecasts = aWeatherForecastList ??
+        throw new System.ArgumentNullException(nameof(aWeatherForecastList));
     }
   }
 }
