+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/WeatherForecast/WeatherForecastState.cs
@@ -1,7 +1,6 @@
 ﻿namespace BlazorHosted_CSharp.Client.Features.WeatherForecast
 {
   using System.Collections.Generic;
-  using System.Reflection;
   using BlazorHosted_CSharp.Shared.Features.WeatherForecast;
   using BlazorState;
 
@@ -14,25 +13,6 @@
     public WeatherForecastsState()
     {
       _WeatherForecasts = new List<WeatherForecastDto>();
-    }
-
-    protected WeatherForecastsState(WeatherForecastsState aState) : this()
-    {
-      foreach (WeatherForecastDto weatherForecastDto in aState.WeatherForecasts)
-      {
-        _WeatherForecasts.Add(weatherForecastDto.Clone() as WeatherForecastDto);
-      }
-    }
-
-    public override object Clone() => new WeatherForecastsState(this);
-
-    protected override void Initialize() { }
-
-    private void Initialize(List<WeatherForecastDto> aWeatherForecastList)
-    {
-      ThrowIfNotTestAssembly(Assembly.GetCallingAssembly());
-      _WeatherForecasts = aWeatherForecastList ??
-        throw new System.ArgumentNullException(nameof(aWeatherForecastList));
     }
   }
 }
